Renamed "EthereumMempool" component -> "EthereumBlockchain"
The `ethereum_txpool` data is formatted a bit differently, and will have
to be handled with different logic.
