+++ frontend/src/components/StreamEntry.js
@@ -2,7 +2,7 @@ import React, { useContext } from "react";
 import { Flex, IconButton, Stack, Tooltip, chakra } from "@chakra-ui/react";
 import { ArrowRightIcon } from "@chakra-ui/icons";
 import UIContext from "../core/providers/UIProvider/context";
-import EthereumMempoolCard from "./stream-cards/EthereumMempool";
+import EthereumBlockchainCard from "./stream-cards/EthereumBlockchain";
 import EthereumWhalewatchCard from "./stream-cards/EthereumWhalewatch";
 
 const StreamEntry_ = ({ entry, showOnboardingTooltips, className }) => {
@@ -36,7 +36,7 @@ const StreamEntry_ = ({ entry, showOnboardingTooltips, className }) => {
         spacing={0}
       >
         {entry.event_type === "ethereum_blockchain" && (
-          <EthereumMempoolCard entry={entry} />
+          <EthereumBlockchainCard entry={entry} />
         )}
 
         {entry.event_type === "ethereum_whalewatch" && (
